refactor(graduate) :  학생 추가 api 훅 이름 변경
