+++ apps/graduate/src/widgets/StudentAddModal/hooks/useSubmitGraduationUser.ts
@@ -15,7 +15,7 @@ export const useGraduationUserSubmit = ({ onSuccess }: Options = {}) => {
 
   const handleAddStudent = async (values: GraduationUserCreateRequest) => {
     await singleMutation.mutateAsync(values);
-    message.success('\ud559\uc0dd\uc774 \ucd94\uac00\ub418\uc5c8\uc2b5\ub2c8\ub2e4.');
+    message.success(' 한 명의 학생이 추가되었습니다.');
     await onSuccess?.();
   };
 
@@ -23,12 +23,12 @@ export const useGraduationUserSubmit = ({ onSuccess }: Options = {}) => {
     rows: GraduationUserCreateRequest[],
   ) => {
     if (!rows.length) {
-      message.warning('\ucd94\uac00\ud560 \ud559\uc0dd\uc774 \uc5c6\uc2b5\ub2c8\ub2e4.');
+      message.warning('추가할 학생이 없습니다.');
       return;
     }
 
     await batchMutation.mutateAsync({ graduationUsers: rows });
-    message.success(`${rows.length}\uba85\uc758 \ud559\uc0dd\uc774 \ucd94\uac00\ub418\uc5c8\uc2b5\ub2c8\ub2e4.`);
+    message.success('총 ' +`${rows.length}` + ' 명의 학생이 추가되었습니다.');
     await onSuccess?.();
   };
